Addressing Jerry's comments. Added more comments to describe behaviors of new filters
