+++ graphjet-core/src/test/java/com/twitter/graphjet/bipartite/segment/MockEdgeTypeMask.java
@@ -1,0 +1,45 @@
+/**
+ * Copyright 2017 Twitter. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.twitter.graphjet.bipartite.segment;
+
+import com.twitter.graphjet.bipartite.api.EdgeTypeMask;
+
+/**
+ * Mocks a certain edge type.
+ */
+public class MockEdgeTypeMask implements EdgeTypeMask {
+  private final byte mockEdgeType;
+
+  public MockEdgeTypeMask(byte mockEdgeType) {
+    this.mockEdgeType = mockEdgeType;
+  }
+
+  @Override
+  public int encode(int node, byte edgeType) {
+    return node;
+  }
+
+  @Override
+  public byte edgeType(int node) {
+    return mockEdgeType;
+  }
+
+  @Override
+  public int restore(int node) {
+    return node;
+  }
+}
